Moved call function to its own file, started fetch function
